% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertir_F_a_C.R
\name{convertir_F_a_C}
\alias{convertir_F_a_C}
\title{Convertir de fahrenheit a celsius}
\usage{
convertir_F_a_C(temp_fahrenheit)
}
\arguments{
\item{temp_fahrenheit}{un vector numérico que represente una temperatura en grados fahrenheit}
}
\value{
un vector numérico de la temperatura ingresada en celsius
}
\description{
Esta función toma una temperatura en grados fahrenheit y la convierte en celsius
}
\examples{
#convertir_F_a_C(100)
#convertir_F_a_C(53.7)

}
