% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suma.R
\name{suma}
\alias{suma}
\title{Suma dos numeros}
\usage{
suma(x = 2, y = 2)
}
\arguments{
\item{x}{un vector numérico}

\item{y}{un vector numérico}
}
\value{
un vector numérico con la suma de \code{x} e \code{y}
}
\description{
La funcion \code{suma()} toma 2 numeros positivos y los suma
}
\details{
Estos son detalles de la función
}
\examples{
suma(2,2)
sum(15,7)

}
